==> atlasGBM/GBMonly/integrate/seurat/36nonNormalSeuratGene2000/code/00_getcountnorm.R <==
library(Matrix)
cell <- names(readRDS('/home-4/whou10@jhu.edu/scratch/Wenpin/brain/atlasGBM/GBMonly/integrate/seurat/36nonNormalSeuratGene2000/res/cluster.rds'))
d <- readRDS('/home-4/whou10@jhu.edu/scratch/Wenpin/brain/atlasGBM/GBMonly/data/36nonNormal_combined_count_mat.rds')
d <- d[,cell]
d <- d[rowSums(d) > 0,]
saveRDS(d,file='/home-4/whou10@jhu.edu/scratch/Wenpin/brain/atlasGBM/GBMonly/integrate/seurat/36nonNormalSeuratGene2000/res/count.rds')
rm('d')
d <- readRDS('/home-4/whou10@jhu.edu/scratch/Wenpin/brain/atlasGBM/GBMonly/data/36nonNormal_combined_log2norm_mat.rds')
d <- d[,cell]
d <- d[rowSums(d) > 0,]
saveRDS(d,file='/home-4/whou10@jhu.edu/scratch/Wenpin/brain/atlasGBM/GBMonly/integrate/seurat/36nonNormalSeuratGene2000/res/norm.rds')
